src/test/java/com/sk/start/start/service/MemberServiceTest.java
package com.sk.start.start.service;

import static org.junit.jupiter.api.Assertions.assertThrows;

import com.sk.start.start.domain.Member;
import com.sk.start.start.repository.MemberRepository;
import com.sk.start.start.repository.MemoryMemberRepository;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

public class MemberServiceTest {
    MemberService memberService;
    MemberRepository memberRepository;

    @BeforeEach
    public void beforeEach() {
        memberRepository = new MemoryMemberRepository();
        memberService = new MemberService(memberRepository);
    }

    @AfterEach
    public void afterEach() {
        memberRepository.clearStore();
    }

    @Test
    public void 회원가입() throws Exception {
        // given
        Member member = new Member();
        member.setName("Hello!");

        // when
        Long saveId = memberService.join(member);

        // then
        Member findMember = memberRepository.findById(saveId).get();
        Assertions.assertThat(member.getName()).isEqualTo(findMember.getName());
    }

    @Test
    public void 중복_회원_예외처리() throws Exception {
        // given
        Member member1 = new Member();
        member1.setName("spring");
        Member member2 = new Member();
        member2.setName("spring");

        // when
        memberService.join(member1);
        
        // then
        IllegalStateException ex = assertThrows(IllegalStateException.class, () -> memberService.join(member2));
        Assertions.assertThat(ex.getMessage()).isEqualTo("기존에 중복된 이름이 있습니다");
    }
}
